SellerService and basic FindAll

+++ SalesWebMVC/Migrations/20210117183828_OtherEntities.cs
@@ -17,7 +17,7 @@ namespace SalesWebMVC.Migrations
                     Name = table.Column<string>(nullable: true),
                     Email = table.Column<string>(nullable: true),
                     BirthDate = table.Column<DateTime>(nullable: false),
-                    BaseSalary = table.Column<decimal>(nullable: false),
+                    BaseSalary = table.Column<double>(nullable: false),
                     DepartmentId = table.Column<int>(nullable: true)
                 },
                 constraints: table =>
@@ -38,7 +38,7 @@ namespace SalesWebMVC.Migrations
                     SalesRecordId = table.Column<int>(nullable: false)
                         .Annotation("MySql:ValueGenerationStrategy", MySqlValueGenerationStrategy.IdentityColumn),
                     Date = table.Column<DateTime>(nullable: false),
-                    Amount = table.Column<decimal>(nullable: false),
+                    Amount = table.Column<double>(nullable: false),
                     Status = table.Column<int>(nullable: false),
                     SellerId = table.Column<int>(nullable: true)
                 },
